+++ src/clj/r99c/routes/home.clj
@@ -16,7 +16,7 @@
    [selmer.filters :refer [add-filter!]]
    [taoensso.timbre :as timbre]))
 
-(timbre/set-level! :debug)
+(timbre/set-level! :info)
 
 (defn- to-date-str [s]
   (-> (str s)