+++ CHANGELOG.md
@@ -13,6 +13,7 @@ int func_test(void) {
 - comments all に pager の導入
 - chart y 軸を動的に。
 
+
 ## 0.14.6 - 2021-12-20
 ### Added
 - display 'enabled indent check' on login page.